+++ lib/field_struct/types/nested.rb
@@ -75,7 +75,7 @@ module FieldStruct
       # @return [String]
       def inspect
         name = @struct_class.name || 'AnonymousFieldStruct'
-        "#<FieldStruct::Types::Nested struct_class=#{name}>"
+        "#<#{FieldStruct.inspect_namespace}::Types::Nested struct_class=#{name}>"
       end
     end
   end